// desktop/src/open-api/model/getRecognitionTasksResponse.ts
/**
 * Receipt Wrangler API.
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RecognitionTask } from './recognitionTask';


export interface GetRecognitionTasksResponse { 
    data: Array<RecognitionTask>;
    totalCount: number;
    activeCount: number;
    awaitingUploadCount: number;
    runningCount: number;
    failedCount: number;
}